Create index, css and js documents
The index archive was created, in which the javascript and the
css documents are called to work in the environment.

The css document aims at changing only the background to dark, so the
white background would not hurt the eyes.

In the javascript document the function that emulates the machine's
election was declared 'computerPLay', in which random values ranging
from 1 to 3 will match either rock, paper or scissors.