Create a function getRandomInteger
In light of the fact that computerPlay() was carrying out two actions,
it was decided to create a helper function to compute a random integer
so that computerPlay would only have to assign Rock, Paper or Scissors
according to the value yielded by getRandomInteger.

The getRandomInteger function leverages the functions Math.floor and
Math.random. The former rounds a number DOWNWARDS to the nearest
integer, whereas the latter computes a number between [0,1), so it had
to be multiplied by 3 to achieve the desired range.

On the other hand, the range was changed to [0,2]; therefore, the
matches would be as follows:

0 - Rock
1 - Paper
2 - Scissors