Create playerElection function
Before creating the singleRound function, which will evaluate the
machine's play against the user's plays, the playerElection function was
created, which will convert the first letter of the player's election
into upper case and the rest of the body into lower case. Thus,
the programm is not sensitive in that aspect.