Create game() function
This function ensures that a game of 5 matches take place. A for loop is
used to achieve this, in which a counter for the machine's points is set
to 0 as well as another one for the user's points.

The comparison to decide who wins every match was possible by slicing
the string displaying whether the user loses or wins, where those were
the key words for comparison.

At the end of the function the points are counted and the winner or a
tie is decided.