Fix game() function and update singleRound()
In order to keep trace of the points of the user and the machine, two
global variables were set to 0. This variables are the default values
for the game() function, which oversees the addition of points for
every player and updates the counters.

The singleRound() function was modified to make the program work. Not
only it returns a String saying if the player won, lost o tied
against the machine, but it also calls the game() function, whose
argument receives the result of the match.

Likewise, the game() function displays the number of points in the UI.