Create transition to point out the winner
In the css archive a new class was added containing the effects to
indicate who the winner is by highlighting either the box of the player
or the machine.

When the counter reaches 5, the corresponding box will become gold.
Furthermore, this action will also stop the counter from increasing
beyond 5.