Style buttons and fix counter
The decideWinner() function was created to reduce lines of code in the
game() function. The decideWinner() oversees when one of the two
containers of the score should illuminate according to the winner.

Likewise, in order to stop the counter from increasing when there is a
winner, if conditions were declared at the beginning of this function to
stop if there is already a winner.

The buttons to choose Rock, Paper or Scissors illuminate when pressed
by creating a new class in the CSS archive, which is added to the botton
that is pressed thanks to event listeners. This effect is erased by
means of the function removeTransition().