Add election of each player in the score
One subcontainer extra was created inside the containers displaying
the scores and the name of the player.

On the other hand, when it was a draw there was no score nor election
in the window. Therefore, the logic for when a draw arises was declared
above the others, so it always runs, but if it is not a draw, it is
uptaded as soon as the program continues.